WIP: search logic is broken
