+++ .gitignore
@@ -16,3 +16,8 @@ node_modules/
 
 # dotenv environment variables file
 .env
+
+# output
+lib/
+
+.DS_STORE
