+++ dict/dataProcessor.js
@@ -1,0 +1,7 @@
+// read data.json
+
+// process data.json based on cluster.js rule
+
+import dict from "./data.json";
+import cluster from "./cluster";
+import pinyin from "pinyin";
