+++ package.json
@@ -10,9 +10,11 @@
   "license": "MIT",
   "main": "index.js",
   "scripts": {
-    "start": "node index.js",
+    "start": "npx babel index.js",
+    "build": "rm -rf lib && babel src -d lib",
+    "exec": "npm run build && node lib/index.js",
+    "clean": "rm -rf lib",
     "test": "mocha",
-    "prettier": "prettier --single-quote --trailing-comma=all --use-tabs --write ./**/*.js",
     "lint": "eslint . --ext js"
   },
   "dependencies": {
@@ -22,6 +24,10 @@
     "request": "^2.88.0"
   },
   "devDependencies": {
+    "@babel/cli": "^7.10.5",
+    "@babel/core": "^7.10.5",
+    "@babel/preset-env": "^7.10.4",
+    "babel-plugin-root-import": "^6.5.0",
     "eslint": "^6.3.0",
     "eslint-config-google": "^0.14.0",
     "eslint-config-standard": "^12.0.0",
@@ -29,8 +35,8 @@
     "eslint-plugin-node": "^8.0.0",
     "eslint-plugin-promise": "^4.2.1",
     "eslint-plugin-standard": "^4.0.1",
-    "mocha": "^5.2.0",
-    "rewire": "^4.0.1",
+    "mocha": "^8.0.1",
+    "rewire": "^5.0.0",
     "standard": "^12.0.1"
   },
   "standard": {
